requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.4
django-extensions==3.2.3
djangorestframework==3.15.1
executing==2.0.1
idna==3.6
iniconfig==2.0.0
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.6
model-bakery==1.17.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pluggy==1.4.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytest==8.1.1
pytest-django==4.8.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
urllib3==2.2.1
wcwidth==0.2.13
